+++ spec/unpoly/fragment_spec.js
@@ -12441,6 +12441,25 @@ describe('up.fragment', function() {
         fixture('.target[up-source="/my-source"]')
         expect(up.fragment.source('.target')).toMatchURL('/my-source')
       })
+
+      it('already reflects the latest source when called from a compiler function', async function() {
+        let sourceSpy = jasmine.createSpy('source spy')
+
+        up.compiler('#two', function(element) {
+          sourceSpy(up.fragment.source(element))
+        })
+
+        fixture('.element#one')
+
+        up.render({ target: '.element', history: true, url: '/url-after-render' })
+        await wait()
+
+        jasmine.respondWithSelector('.element#two')
+        await wait()
+
+        expect(sourceSpy).toHaveBeenCalledWith('/url-after-render')
+      })
+
     })
 
     describe('up.fragment.failKey', function() {
@@ -13869,6 +13888,38 @@ describe('up.fragment', function() {
       })
     })
 
+    describe('up.fragment.defaultNormalizeKeepHTML()', function() {
+
+      for (let attr of ['nonce', 'up-etag', 'up-time', 'up-source', 'up-on-rendered', 'up-watch', 'up-on-accepted']) {
+
+        it(`removes a [${attr}] attribute from the root element`, function() {
+          let input = `<div ${attr}="value"></div>`
+          let output = up.fragment.defaultNormalizeKeepHTML(input)
+          expect(output).toBe('<div></div>')
+        })
+
+        it(`removes a [${attr}] attribute but leaves other attributes`, function() {
+          let input = `<div data-x="x" ${attr}="value" data-y="y"></div>`
+          let output = up.fragment.defaultNormalizeKeepHTML(input)
+          expect(output).toBe('<div data-x="x" data-y="y"></div>')
+        })
+
+        it(`removes multiple [${attr}] attributes in descendants`, function() {
+          let input = `<div><span ${attr}="value"></span><span ${attr}="value"></span></div>`
+          let output = up.fragment.defaultNormalizeKeepHTML(input)
+          expect(output).toBe('<div><span></span><span></span></div>')
+        })
+
+      }
+
+      it('removes multiple attributes from the same tag', function() {
+        let input = `<a id="keepable" href="/path" up-keep="same-html" up-on-rendered="nonce-page-secret foo()" data-x="x" up-on-finished="nonce-page-secret bar()" data-y="y">text</a>`
+        let output = up.fragment.defaultNormalizeKeepHTML(input)
+        expect(output).toBe(`<a id="keepable" href="/path" up-keep="same-html" data-x="x" data-y="y">text</a>`)
+      })
+
+    })
+
   })
 
   describe('unobtrusive behavior', function() {
@@ -13885,7 +13936,7 @@ describe('up.fragment', function() {
       }
 
       beforeEach(function() {
-        // Need to refactor this spec file so examples don't all share one example
+        // TODO: Refactor this spec file so examples don't all share one example
         $('.before, .middle, .after').remove()
       })
 
@@ -14886,6 +14937,307 @@ describe('up.fragment', function() {
         expect(oldTextDuringTransition).toEqual('old-foo old-bar')
         expect(newTextDuringTransition).toEqual('new-foo old-bar')
       })
+
+      function keepableFixture(html, renderOptions = {}) {
+        // Make sure we insert the element through rendering, so we can test
+        // our own HTML pollution with [up-etag], [up-time], [up-source].
+        fixture('#keepable')
+        up.render({ target: '#keepable', document: html, ...renderOptions })
+
+        return document.querySelector('#keepable')
+      }
+
+      describe('with [up-keep="same-html"]', function() {
+
+        it('keeps the element if its outer HTML is stable', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-html">text</div>')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-html">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('keeps an [up-keep] descendant if its outer HTML is stable', async function() {
+          let [container, keepable] = htmlFixtureList(`
+            <div id="container">
+              <div id="keepable" up-keep="same-html">text</div>
+            </div>
+          `)
+          up.hello(container)
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-html">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('ignores indentation', async function() {
+          let [container, keepable] = htmlFixtureList(`
+            <div id="container">
+              <div id="keepable" up-keep="same-html">
+                text
+              </div>
+            </div>
+          `)
+          up.hello(container)
+
+          up.render({ fragment: `
+                  <div id="keepable" up-keep="same-html">
+                          text
+                  </div>
+          ` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('ignores HTML changes made by a user compiler', async function() {
+          up.compiler('#keepable', (el) => el.innerText = 'text changed by compiler')
+
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-html">text</div>')
+          expect(keepable).toHaveText('text changed by compiler')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-html">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+          expect(document.querySelector('#keepable')).toHaveText('text changed by compiler')
+        })
+
+        it('ignores HTML changes made by a user macro', async function() {
+          up.macro('#keepable', (el) => el.innerText = 'text changed by macro')
+
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-html">text</div>')
+          expect(keepable).toHaveText('text changed by macro')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-html">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+          expect(document.querySelector('#keepable')).toHaveText('text changed by macro')
+        })
+
+        it('ignores HTML changes made by a system macro', async function() {
+          html = `
+            <div id="keepable" up-expand up-keep="same-html">
+              <a href="/path">text</a>
+            </div>
+          `
+          let keepable = keepableFixture(html)
+          expect(keepable).toBeFollowable()
+          expect(keepable).toHaveAttribute('up-href', '/path')
+
+          up.render({ fragment: html })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('keeps an identical element that was introduced by a transition (which adds tracking classes)', async function() {
+          fixture('#keepable')
+          html = '<div id="keepable" up-keep="same-html">text</div>'
+          let keepable = (await up.render({ target: '#keepable', document: html, transition: 'cross-fade', duration: 70 }).finished).fragment
+
+          up.render({ fragment: html })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('keeps an element when a new identitcal was introduced by a transition (which adds tracking classes)', async function() {
+          html = '<div id="keepable" up-keep="same-html">text</div>'
+          let keepable = keepableFixture(html)
+
+          await up.render({ fragment: html, transition: 'cross-fade', duration: 70 }).finished
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('ignores HTML changes that happened after rendering', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-html">text</div>')
+
+          keepable.innerText = 'text changed after rendering'
+          expect(keepable).toHaveText('text changed after rendering')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-html">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+          expect(document.querySelector('#keepable')).toHaveText('text changed after rendering')
+        })
+
+        it('replaces the element if its text changed', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-html">old text</div>')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-html">new text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).not.toBe(keepable)
+        })
+
+        it('replaces the element if an attribute changed', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-html" attr="old">text</div>')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-html" attr="new">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).not.toBe(keepable)
+        })
+
+        it('replaces the element if child element changed', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-html" attr="old"></div>')
+
+          up.render({ fragment: `
+            <div id="keepable" up-keep="same-html" attr="new">
+              <span></span>
+            </div>
+          `})
+          await wait()
+
+          expect(document.querySelector('#keepable')).not.toBe(keepable)
+        })
+
+        it('ignores a <script nonce> for the comparison', async function() {
+          up.fragment.config.runScripts = true
+
+          spyOn(up.protocol, 'cspNonce').and.returnValue('page-secret')
+
+          let keepable = keepableFixture(`<div id="keepable" up-keep="same-html"><script nonce="page-secret"></script></div>`)
+
+          up.render({ target: '#keepable', url: '/path' })
+          await wait()
+
+          jasmine.respondWith({
+            responseText: `<div id="keepable" up-keep="same-html"><script nonce="response-secret"></script></div>`,
+            responseHeaders: { 'Content-Security-Policy': "script-src: 'nonce-response-secret'" }
+          })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('ignores a callback nonce for the comparison', async function() {
+          up.fragment.config.runScripts = true
+
+          spyOn(up.protocol, 'cspNonce').and.returnValue('page-secret')
+
+          let keepable = keepableFixture(`<input id="keepable" name="foo" up-keep="same-html" up-watch="nonce-page-secret foo()"></input>`)
+
+          up.render({ target: '#keepable', url: '/path' })
+          await wait()
+
+          jasmine.respondWith({
+            responseText: `<input id="keepable" name="foo" up-keep="same-html" up-watch="nonce-response-secret foo()"></input>`,
+            responseHeaders: { 'Content-Security-Policy': "script-src: 'nonce-response-secret'" }
+          })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('ignores multiple callback nonces within the same tag', async function() {
+          up.fragment.config.runScripts = true
+
+          spyOn(up.protocol, 'cspNonce').and.returnValue('page-secret')
+
+          let keepable = keepableFixture(`<a id="keepable" href="/path" up-keep="same-html" up-on-rendered="nonce-page-secret foo()" data-x="x" up-on-finished="nonce-page-secret bar()" data-y="y">text</a>`)
+
+          up.render({ target: '#keepable', url: '/path' })
+          await wait()
+
+          jasmine.respondWith({
+            responseText: `<a id="keepable" href="/path" up-keep="same-html" up-on-rendered="nonce-response-secret foo()" data-x="x" up-on-finished="nonce-response-secret bar()" data-y="y">text</a>`,
+            responseHeaders: { 'Content-Security-Policy': "script-src: 'nonce-response-secret'" }
+          })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+      })
+
+      describe('with [up-keep="same-data"]', function() {
+
+        it('keeps the element if its data is stable, even though the innerHTML changed', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">text</div>')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">changed text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('keeps an [up-keep] descendant if its data is stable, even though the innerHTML changed', async function() {
+          let [container, keepable] = htmlFixtureList(`
+            <div id="container">
+              <div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">text</div>
+            </div>
+          `)
+          up.hello(container)
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">changed text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('keeps an element with an empty [up-data] object when the new element has no [up-data] attribute at all')
+
+        it('merges [up-data] and [data-*] attribute in both elements before comparison')
+
+        it('ignores data mutated after rendering', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">text</div>')
+
+          up.data(keepable).bar = 2
+          expect(up.data(keepable)).toEqual({ foo: 1, bar: 2})
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).toBe(keepable)
+        })
+
+        it('replaces the element if a data value changed', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">text</div>')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-data" up-data="{ foo: 2 }">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).not.toBe(keepable)
+        })
+
+        it('replaces the element if the new element has an additional data key', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">text</div>')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-data" up-data="{ foo: 1, bar: 2 }">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).not.toBe(keepable)
+        })
+
+        it('replaces the element if has data, but the new element does not', async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-data" up-data="{ foo: 1 }">text</div>')
+
+          up.render({ fragment: `<div id="keepable" up-keep="same-data">text</div>` })
+          await wait()
+
+          expect(document.querySelector('#keepable')).not.toBe(keepable)
+        })
+
+        it("replaces the element if it doesn't have data, but the new element has data", async function() {
+          let keepable = keepableFixture('<div id="keepable" up-keep="same-data"></div>')
+
+          up.render({ fragment: `
+            <div id="keepable" up-keep="same-data" up-data="{ foo: 1 }"></div>
+          `})
+          await wait()
+
+          expect(document.querySelector('#keepable')).not.toBe(keepable)
+        })
+
+      })
+
     })
 
   })