+++ src/unpoly/classes/change/open_layer.js
@@ -130,11 +130,15 @@ up.Change.OpenLayer = class OpenLayer extends up.Change.Addition {
     // on background layers.
     this.layer.setContent(this._content)
 
+    // Adopt CSP nonces and fix broken script tags
+    this.responseDoc.finalizeElement(this._content)
+
     // Remember where the element came from to support up.reload(element).
     this.setReloadAttrs({ newElement: this._content, source: this.options.source })
 
-    // Adopt CSP nonces and fix broken script tags
-    this.responseDoc.finalizeElement(this._content)
+    // Compile the entire layer, not just the user content.
+    // E.g. [up-dismiss] in the layer elements needs to go through a macro.
+    up.hello(this.layer.element, { ...this.options, layer: this.layer, dataRoot: this._content })
 
     this._newOverlayResult = new up.RenderResult({
       layer: this.layer,
@@ -142,10 +146,6 @@ up.Change.OpenLayer = class OpenLayer extends up.Change.Addition {
       target: this.target,
       renderOptions: this.options,
     })
-
-    // Compile the entire layer, not just the user content.
-    // E.g. [up-dismiss] in the layer elements needs to go through a macro.
-    up.hello(this.layer.element, { ...this.options, layer: this.layer, dataRoot: this._content })
 
     // Don't wait for the open animation to finish.
     // Otherwise a popup would start to open and only reveal itself after the animation.